Rename `topology` module to `objects`
The new name doesn't quite fit yet, but I plan to move all kinds of
objects in there, not just topological ones.
