+++ crates/fj-kernel/src/lib.rs
@@ -90,5 +90,5 @@
 pub mod algorithms;
 pub mod builder;
 pub mod geometry;
+pub mod objects;
 pub mod shape;
-pub mod topology;
