+++ crates/fj-kernel/src/objects/cycle.rs
