+++ crates/fj-kernel/src/shape/validate/mod.rs
@@ -12,7 +12,7 @@ use fj_math::Scalar;
 
 use crate::{
     geometry::{Curve, Surface},
-    topology::{Cycle, Edge, Face, Vertex},
+    objects::{Cycle, Edge, Face, Vertex},
 };
 
 use super::{stores::Stores, Handle, Object};
