+++ crates/fj-operations/src/difference_2d.rs
@@ -1,8 +1,8 @@
 use fj_interop::debug::DebugInfo;
 use fj_kernel::{
     algorithms::Tolerance,
+    objects::{Cycle, Edge, Face},
     shape::{LocalForm, Shape, ValidationError},
-    topology::{Cycle, Edge, Face},
 };
 use fj_math::Aabb;
