+++ crates/fj-operations/src/sketch.rs
@@ -2,8 +2,8 @@ use fj_interop::debug::DebugInfo;
 use fj_kernel::{
     algorithms::Tolerance,
     geometry::Surface,
+    objects::Face,
     shape::{Shape, ValidationError},
-    topology::Face,
 };
 use fj_math::{Aabb, Point};
